initial upload
This is the first upload of files from a previous class project. We had to make a functioning site with one page having "piazza". I made a fish tank. Will be removing the extra pages that is not tank page. Will also be redoing the code so it is with Vanilla javascript and more of my own custom CSS; currently uses materialize and JQuery.